Remove the static limit of 3 items in the ActionBar
The limit is now set in the theme. The default value is still 3 as this is considered as the maximum number of global actions a screen SHOULD have. Please do not change this value unless you are completely aware of what you are doing. Modifying it may impair user experience. If you really have a large number of actions, please consider QuickActions.

Closes #36